+++ packages/lib/dental/index.ts
@@ -15,6 +15,8 @@ export * from "./two-factor-policy";
 export * from "./assert-two-factor-enabled";
 export * from "./pvs/enqueue-pvs-sync";
 export * from "./pvs/pvs-outbox.constants";
+export * from "./pvs/pvs-connector-key";
+export * from "./pvs/pvs-connector-credential.service";
 export * from "./pvs/pvs-connector-auth";
 export * from "./pvs/pvs-outbox.service";
 export * from "./smart-fill";